+++ extensions-leaflet-map.php
@@ -4,7 +4,7 @@
  * Description:       Extends the WordPress Plugin <a href="https://wordpress.org/plugins/leaflet-map/">Leaflet Map</a> with Leaflet Plugins and other functions.
  * Plugin URI:        https://leafext.de/en/
  * Update URI:        https://github.com/hupe13/extensions-leaflet-map-github
- * Version:           4.5-250809
+ * Version:           4.5-250812
  * Requires PHP:      7.4
  * Requires Plugins:  leaflet-map
  * Author:            hupe13
@@ -37,8 +37,8 @@ if ( ! function_exists( 'leafext_plugin_active' ) ) {
 		if ( count( $is_active ) === 1 ) {
 			return true;
 		}
-		$plugins   = get_site_option( 'active_sitewide_plugins' );
-		if (is_array($plugins)) {
+		$plugins = get_site_option( 'active_sitewide_plugins' );
+		if ( is_array( $plugins ) ) {
 			$is_active = preg_grep( '/^.*\/' . $slug . '\.php$/', array_flip( $plugins ) );
 			if ( count( $is_active ) === 1 ) {
 				return true;
@@ -111,22 +111,25 @@ function leafext_extra_textdomain() {
 add_action( 'plugins_loaded', 'leafext_extra_textdomain' );
 
 // WP < 6.5, ClassicPress
-function leafext_leaflet_require() {
-	if ( ! leafext_plugin_active( 'leaflet-map' ) ) {
-		deactivate_plugins( plugin_basename( __FILE__ ) );
-		$message = '<div><p>' . sprintf(
+if ( function_exists( 'classicpress_version' ) ||
+( function_exists( 'wp_get_wp_version' ) && version_compare( '6.5', wp_get_wp_version(), '>' ) ) ) {
+	function leafext_leaflet_require() {
+		if ( ! leafext_plugin_active( 'leaflet-map' ) ) {
+			deactivate_plugins( plugin_basename( __FILE__ ) );
+			$message = '<div><p>' . sprintf(
 			/* translators: %s are plugin names. */
-			esc_html__( 'Please install and activate %1$s before using %2$s.', 'extensions-leaflet-map' ),
-			'<a href="https://wordpress.org/plugins/leaflet-map/">Leaflet Map</a>',
-			'Extensions for Leaflet Map'
-		) . '</p><p><a href="' . esc_html( network_admin_url( 'plugins.php' ) ) . '">' .
-			__( 'Manage plugins', 'extensions-leaflet-map' ) . '</a>.</p></div>';
-		$error = new WP_Error( 'error', $message );
-		// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- it is an WP Error
-		wp_die( $error, '', wp_kses_post( $error->get_error_data() ) );
+				esc_html__( 'Please install and activate %1$s before using %2$s.', 'extensions-leaflet-map' ),
+				'<a href="https://wordpress.org/plugins/leaflet-map/">Leaflet Map</a>',
+				'Extensions for Leaflet Map'
+			) . '</p><p><a href="' . esc_html( network_admin_url( 'plugins.php' ) ) . '">' .
+				__( 'Manage plugins', 'extensions-leaflet-map' ) . '</a>.</p></div>';
+			$error = new WP_Error( 'error', $message );
+			// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- it is an WP Error
+			wp_die( $error, '', wp_kses_post( $error->get_error_data() ) );
+		}
 	}
+	register_activation_hook( __FILE__, 'leafext_leaflet_require' );
 }
-register_activation_hook( __FILE__, 'leafext_leaflet_require' );
 
 // Disable activation the other of WP / Github Version
 if ( ! function_exists( 'leafext_disable_extensions_activation' ) ) {