+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * fixed: leafext_plugin_active did not work with network activated plugins
 * new Jshrink
 * new leaflet.fullscreen
+* zoomhomemap has some options now
 
 ### 4.5.1 / 250604